chore(tests): Restructured adder tests.
